feat: gate foundry-primitives behind optimism feature
