+++ crates/anvil/Cargo.toml
@@ -28,7 +28,7 @@ foundry-common.workspace = true
 foundry-config.workspace = true
 foundry-evm.workspace = true
 foundry-evm-networks.workspace = true
-foundry-primitives.workspace = true
+foundry-primitives = { workspace = true, features = ["optimism"] }
 tempo-chainspec.workspace = true
 tempo-primitives.workspace = true
 tempo-precompiles.workspace = true
