+++ crates/anvil/core/Cargo.toml
@@ -15,7 +15,7 @@ workspace = true
 [dependencies]
 foundry-common.workspace = true
 foundry-evm.workspace = true
-foundry-primitives.workspace = true
+foundry-primitives = { workspace = true, features = ["optimism"] }
 revm = { workspace = true, default-features = false, features = [
     "std",
     "serde",
