+++ crates/primitives/src/network/mod.rs
@@ -1,11 +1,19 @@
 use alloy_network::Network;
 
+#[cfg(feature = "optimism")]
+mod optimism;
 mod receipt;
 
 use alloy_provider::fillers::{
     BlobGasFiller, ChainIdFiller, GasFiller, JoinFill, NonceFiller, RecommendedFillers,
 };
+#[cfg(feature = "optimism")]
+pub use optimism::FoundryTransactionResponse;
 pub use receipt::*;
+
+/// Default JSON-RPC transaction response when the `optimism` feature is disabled.
+#[cfg(not(feature = "optimism"))]
+pub type FoundryTransactionResponse = alloy_rpc_types_eth::Transaction<crate::FoundryTxEnvelope>;
 
 /// Foundry network type.
 ///
@@ -36,7 +44,7 @@ impl Network for FoundryNetwork {
 
     type TransactionRequest = crate::FoundryTransactionRequest;
 
-    type TransactionResponse = op_alloy_rpc_types::Transaction<crate::FoundryTxEnvelope>;
+    type TransactionResponse = FoundryTransactionResponse;
 
     type ReceiptResponse = crate::FoundryTxReceipt;
