+++ crates/primitives/src/transaction/envelope.rs
@@ -1,6 +1,7 @@
+#[cfg(feature = "optimism")]
+use alloy_consensus::{Sealed, Transaction as _};
 use alloy_consensus::{
-    Sealed, Signed, Transaction as _, TransactionEnvelope, TxEip1559, TxEip2930, TxEnvelope,
-    TxLegacy, TxType, Typed2718,
+    Signed, TransactionEnvelope, TxEip1559, TxEip2930, TxEnvelope, TxLegacy, TxType, Typed2718,
     crypto::RecoveryError,
     transaction::{
         SignerRecoverable, TxEip7702, TxHashRef,
@@ -9,14 +10,10 @@ use alloy_consensus::{
 };
 use alloy_evm::{FromRecoveredTx, FromTxWithEncoded};
 use alloy_network::{AnyRpcTransaction, AnyTxEnvelope, TransactionResponse};
-use alloy_op_evm::OpTx;
 use alloy_primitives::{Address, B256, Bytes, TxHash};
 use alloy_rpc_types::ConversionError;
-use op_alloy_consensus::{
-    DEPOSIT_TX_TYPE_ID, OpTransaction as OpTransactionTrait, POST_EXEC_TX_TYPE_ID, TxDeposit,
-    TxPostExec,
-};
-use op_revm::{OpTransaction, transaction::deposit::DepositTransactionParts};
+#[cfg(feature = "optimism")]
+use op_alloy_consensus::{DEPOSIT_TX_TYPE_ID, POST_EXEC_TX_TYPE_ID, TxDeposit, TxPostExec};
 use revm::context::TxEnv;
 use tempo_primitives::{AASigned, TempoTransaction};
 use tempo_revm::TempoTxEnv;
@@ -57,9 +54,11 @@ pub enum FoundryTxEnvelope {
     /// OP stack deposit transaction.
     ///
     /// See <https://docs.optimism.io/op-stack/bridging/deposit-flow>.
+    #[cfg(feature = "optimism")]
     #[envelope(ty = 126)]
     Deposit(Sealed<TxDeposit>),
     /// OP stack post-execution synthetic transaction.
+    #[cfg(feature = "optimism")]
     #[envelope(ty = 0x7D)]
     PostExec(Sealed<TxPostExec>),
     /// Tempo transaction type.
@@ -80,7 +79,9 @@ impl FoundryTxEnvelope {
             Self::Eip1559(tx) => Ok(TxEnvelope::Eip1559(tx)),
             Self::Eip4844(tx) => Ok(TxEnvelope::Eip4844(tx)),
             Self::Eip7702(tx) => Ok(TxEnvelope::Eip7702(tx)),
+            #[cfg(feature = "optimism")]
             Self::Deposit(_) => Err(self),
+            #[cfg(feature = "optimism")]
             Self::PostExec(_) => Err(self),
             Self::Tempo(_) => Err(self),
         }
@@ -109,7 +110,9 @@ impl FoundryTxEnvelope {
             Self::Eip1559(t) => *t.hash(),
             Self::Eip4844(t) => *t.hash(),
             Self::Eip7702(t) => *t.hash(),
+            #[cfg(feature = "optimism")]
             Self::Deposit(t) => t.tx_hash(),
+            #[cfg(feature = "optimism")]
             Self::PostExec(t) => t.tx_hash(),
             Self::Tempo(t) => *t.hash(),
         }
@@ -128,7 +131,9 @@ impl FoundryTxEnvelope {
             Self::Eip1559(tx) => tx.recover_signer()?,
             Self::Eip4844(tx) => tx.recover_signer()?,
             Self::Eip7702(tx) => tx.recover_signer()?,
+            #[cfg(feature = "optimism")]
             Self::Deposit(tx) => tx.from,
+            #[cfg(feature = "optimism")]
             Self::PostExec(tx) => tx.inner().signer_address(),
             Self::Tempo(tx) => tx.signature().recover_signer(&tx.signature_hash())?,
         })
@@ -143,7 +148,9 @@ impl TxHashRef for FoundryTxEnvelope {
             Self::Eip1559(t) => t.hash(),
             Self::Eip4844(t) => t.hash(),
             Self::Eip7702(t) => t.hash(),
+            #[cfg(feature = "optimism")]
             Self::Deposit(t) => t.hash_ref(),
+            #[cfg(feature = "optimism")]
             Self::PostExec(t) => t.hash_ref(),
             Self::Tempo(t) => t.hash(),
         }
@@ -157,23 +164,6 @@ impl SignerRecoverable for FoundryTxEnvelope {
 
     fn recover_signer_unchecked(&self) -> Result<Address, RecoveryError> {
         self.recover()
-    }
-}
-
-impl OpTransactionTrait for FoundryTxEnvelope {
-    fn is_deposit(&self) -> bool {
-        matches!(self, Self::Deposit(_))
-    }
-
-    fn as_deposit(&self) -> Option<&Sealed<TxDeposit>> {
-        match self {
-            Self::Deposit(tx) => Some(tx),
-            _ => None,
-        }
-    }
-
-    fn as_post_exec(&self) -> Option<&Sealed<TxPostExec>> {
-        if let Self::PostExec(tx) = self { Some(tx) } else { None }
     }
 }
 
@@ -193,19 +183,6 @@ impl From<TxEnvelope> for FoundryTxEnvelope {
             TxEnvelope::Eip1559(tx) => Self::Eip1559(tx),
             TxEnvelope::Eip4844(tx) => Self::Eip4844(tx),
             TxEnvelope::Eip7702(tx) => Self::Eip7702(tx),
-        }
-    }
-}
-
-impl From<op_alloy_consensus::OpTxEnvelope> for FoundryTxEnvelope {
-    fn from(tx: op_alloy_consensus::OpTxEnvelope) -> Self {
-        match tx {
-            op_alloy_consensus::OpTxEnvelope::Legacy(tx) => Self::Legacy(tx),
-            op_alloy_consensus::OpTxEnvelope::Eip2930(tx) => Self::Eip2930(tx),
-            op_alloy_consensus::OpTxEnvelope::Eip1559(tx) => Self::Eip1559(tx),
-            op_alloy_consensus::OpTxEnvelope::Eip7702(tx) => Self::Eip7702(tx),
-            op_alloy_consensus::OpTxEnvelope::Deposit(tx) => Self::Deposit(tx),
-            op_alloy_consensus::OpTxEnvelope::PostExec(tx) => Self::PostExec(tx),
         }
     }
 }
@@ -236,33 +213,50 @@ impl TryFrom<AnyRpcTransaction> for FoundryTxEnvelope {
                 TxEnvelope::Eip4844(tx) => Ok(Self::Eip4844(tx)),
                 TxEnvelope::Eip7702(tx) => Ok(Self::Eip7702(tx)),
             },
-            AnyTxEnvelope::Unknown(mut tx) => {
-                // Try to convert to deposit transaction
-                if tx.ty() == DEPOSIT_TX_TYPE_ID {
-                    tx.inner.fields.insert("from".to_string(), serde_json::to_value(from).unwrap());
-                    let deposit_tx =
-                        tx.inner.fields.deserialize_into::<TxDeposit>().map_err(|e| {
-                            ConversionError::Custom(format!(
-                                "Failed to deserialize deposit tx: {e}"
-                            ))
-                        })?;
+            AnyTxEnvelope::Unknown(tx) => {
+                #[cfg(feature = "optimism")]
+                {
+                    let mut tx = tx;
+                    let _ = from;
+                    // Try to convert to deposit transaction
+                    if tx.ty() == DEPOSIT_TX_TYPE_ID {
+                        tx.inner
+                            .fields
+                            .insert("from".to_string(), serde_json::to_value(from).unwrap());
+                        let deposit_tx =
+                            tx.inner.fields.deserialize_into::<TxDeposit>().map_err(|e| {
+                                ConversionError::Custom(format!(
+                                    "Failed to deserialize deposit tx: {e}"
+                                ))
+                            })?;
 
-                    return Ok(Self::Deposit(Sealed::new(deposit_tx)));
+                        return Ok(Self::Deposit(Sealed::new(deposit_tx)));
+                    }
+
+                    if tx.ty() == POST_EXEC_TX_TYPE_ID {
+                        let post_exec_tx =
+                            tx.inner.fields.deserialize_into::<TxPostExec>().map_err(|e| {
+                                ConversionError::Custom(format!(
+                                    "Failed to deserialize post-exec tx: {e}"
+                                ))
+                            })?;
+
+                        return Ok(Self::PostExec(Sealed::new(post_exec_tx)));
+                    }
+
+                    let tx_type = tx.ty();
+                    Err(ConversionError::Custom(format!(
+                        "Unknown transaction type: 0x{tx_type:02X}"
+                    )))
                 }
-
-                if tx.ty() == POST_EXEC_TX_TYPE_ID {
-                    let post_exec_tx =
-                        tx.inner.fields.deserialize_into::<TxPostExec>().map_err(|e| {
-                            ConversionError::Custom(format!(
-                                "Failed to deserialize post-exec tx: {e}"
-                            ))
-                        })?;
-
-                    return Ok(Self::PostExec(Sealed::new(post_exec_tx)));
+                #[cfg(not(feature = "optimism"))]
+                {
+                    let _ = from;
+                    let tx_type = tx.ty();
+                    Err(ConversionError::Custom(format!(
+                        "Unknown transaction type: 0x{tx_type:02X}"
+                    )))
                 }
-
-                let tx_type = tx.ty();
-                Err(ConversionError::Custom(format!("Unknown transaction type: 0x{tx_type:02X}")))
             }
         }
     }
@@ -276,6 +270,7 @@ impl FromRecoveredTx<FoundryTxEnvelope> for TxEnv {
             FoundryTxEnvelope::Eip1559(signed_tx) => Self::from_recovered_tx(signed_tx, caller),
             FoundryTxEnvelope::Eip4844(signed_tx) => Self::from_recovered_tx(signed_tx, caller),
             FoundryTxEnvelope::Eip7702(signed_tx) => Self::from_recovered_tx(signed_tx, caller),
+            #[cfg(feature = "optimism")]
             FoundryTxEnvelope::Deposit(sealed_tx) => {
                 let tx = sealed_tx.inner();
                 Self {
@@ -288,6 +283,7 @@ impl FromRecoveredTx<FoundryTxEnvelope> for TxEnv {
                     ..Default::default()
                 }
             }
+            #[cfg(feature = "optimism")]
             FoundryTxEnvelope::PostExec(sealed_tx) => {
                 let tx = sealed_tx.inner();
                 Self {
@@ -299,63 +295,6 @@ impl FromRecoveredTx<FoundryTxEnvelope> for TxEnv {
                 }
             }
             FoundryTxEnvelope::Tempo(_) => unreachable!("Tempo tx in Ethereum context"),
-        }
-    }
-}
-
-impl FromRecoveredTx<FoundryTxEnvelope> for OpTransaction<TxEnv> {
-    fn from_recovered_tx(tx: &FoundryTxEnvelope, caller: Address) -> Self {
-        match tx {
-            FoundryTxEnvelope::Legacy(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: None, deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip2930(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: None, deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip1559(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: None, deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip4844(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: None, deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip7702(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: None, deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Deposit(sealed_tx) => {
-                let deposit_tx = sealed_tx.inner();
-                let base = TxEnv {
-                    tx_type: deposit_tx.ty(),
-                    caller,
-                    gas_limit: deposit_tx.gas_limit,
-                    kind: deposit_tx.to,
-                    value: deposit_tx.value,
-                    data: deposit_tx.input.clone(),
-                    ..Default::default()
-                };
-                let deposit = DepositTransactionParts {
-                    source_hash: deposit_tx.source_hash,
-                    mint: Some(deposit_tx.mint),
-                    is_system_transaction: deposit_tx.is_system_transaction,
-                };
-                Self { base, enveloped_tx: None, deposit }
-            }
-            FoundryTxEnvelope::PostExec(sealed_tx) => {
-                let tx = sealed_tx.inner();
-                let base = TxEnv {
-                    tx_type: tx.ty(),
-                    caller,
-                    kind: tx.kind(),
-                    data: tx.input.clone(),
-                    ..Default::default()
-                };
-                Self { base, enveloped_tx: None, deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Tempo(_) => unreachable!("Tempo tx in Optimism context"),
         }
     }
 }
@@ -384,7 +323,9 @@ impl FromRecoveredTx<FoundryTxEnvelope> for TempoTxEnv {
             FoundryTxEnvelope::Eip7702(signed_tx) => {
                 Self::from(TxEnv::from_recovered_tx(signed_tx, caller))
             }
+            #[cfg(feature = "optimism")]
             FoundryTxEnvelope::Deposit(_) => unreachable!("Deposit tx in Tempo context"),
+            #[cfg(feature = "optimism")]
             FoundryTxEnvelope::PostExec(_) => unreachable!("Post-exec tx in Tempo context"),
             FoundryTxEnvelope::Tempo(aa_signed) => Self::from_recovered_tx(aa_signed, caller),
         }
@@ -397,75 +338,6 @@ impl FromTxWithEncoded<FoundryTxEnvelope> for TempoTxEnv {
     }
 }
 
-impl FromRecoveredTx<FoundryTxEnvelope> for OpTx {
-    fn from_recovered_tx(tx: &FoundryTxEnvelope, caller: Address) -> Self {
-        Self(OpTransaction::<TxEnv>::from_recovered_tx(tx, caller))
-    }
-}
-
-impl FromTxWithEncoded<FoundryTxEnvelope> for OpTx {
-    fn from_encoded_tx(tx: &FoundryTxEnvelope, caller: Address, encoded: Bytes) -> Self {
-        Self(OpTransaction::<TxEnv>::from_encoded_tx(tx, caller, encoded))
-    }
-}
-
-impl FromTxWithEncoded<FoundryTxEnvelope> for OpTransaction<TxEnv> {
-    fn from_encoded_tx(tx: &FoundryTxEnvelope, caller: Address, encoded: Bytes) -> Self {
-        match tx {
-            FoundryTxEnvelope::Legacy(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: Some(encoded), deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip2930(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: Some(encoded), deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip1559(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: Some(encoded), deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip4844(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: Some(encoded), deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Eip7702(signed_tx) => {
-                let base = TxEnv::from_recovered_tx(signed_tx, caller);
-                Self { base, enveloped_tx: Some(encoded), deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Deposit(sealed_tx) => {
-                let deposit_tx = sealed_tx.inner();
-                let base = TxEnv {
-                    tx_type: deposit_tx.ty(),
-                    caller,
-                    gas_limit: deposit_tx.gas_limit,
-                    kind: deposit_tx.to,
-                    value: deposit_tx.value,
-                    data: deposit_tx.input.clone(),
-                    ..Default::default()
-                };
-                let deposit = DepositTransactionParts {
-                    source_hash: deposit_tx.source_hash,
-                    mint: Some(deposit_tx.mint),
-                    is_system_transaction: deposit_tx.is_system_transaction,
-                };
-                Self { base, enveloped_tx: Some(encoded), deposit }
-            }
-            FoundryTxEnvelope::PostExec(sealed_tx) => {
-                let tx = sealed_tx.inner();
-                let base = TxEnv {
-                    tx_type: tx.ty(),
-                    caller,
-                    kind: tx.kind(),
-                    data: tx.input.clone(),
-                    ..Default::default()
-                };
-                Self { base, enveloped_tx: Some(encoded), deposit: Default::default() }
-            }
-            FoundryTxEnvelope::Tempo(_) => unreachable!("Tempo tx in Optimism context"),
-        }
-    }
-}
-
 impl std::fmt::Display for FoundryTxType {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         match self {
@@ -474,7 +346,9 @@ impl std::fmt::Display for FoundryTxType {
             Self::Eip1559 => write!(f, "eip1559"),
             Self::Eip4844 => write!(f, "eip4844"),
             Self::Eip7702 => write!(f, "eip7702"),
+            #[cfg(feature = "optimism")]
             Self::Deposit => write!(f, "deposit"),
+            #[cfg(feature = "optimism")]
             Self::PostExec => write!(f, "post-exec"),
             Self::Tempo => write!(f, "tempo"),
         }
@@ -501,7 +375,9 @@ impl From<FoundryTxEnvelope> for FoundryTypedTx {
             FoundryTxEnvelope::Eip1559(signed_tx) => Self::Eip1559(signed_tx.strip_signature()),
             FoundryTxEnvelope::Eip4844(signed_tx) => Self::Eip4844(signed_tx.strip_signature()),
             FoundryTxEnvelope::Eip7702(signed_tx) => Self::Eip7702(signed_tx.strip_signature()),
+            #[cfg(feature = "optimism")]
             FoundryTxEnvelope::Deposit(sealed_tx) => Self::Deposit(sealed_tx.into_inner()),
+            #[cfg(feature = "optimism")]
             FoundryTxEnvelope::PostExec(sealed_tx) => Self::PostExec(sealed_tx.into_inner()),
             FoundryTxEnvelope::Tempo(signed_tx) => Self::Tempo(signed_tx.strip_signature()),
         }
@@ -610,28 +486,6 @@ mod tests {
     }
 
     #[test]
-    fn test_decode_encode_deposit_tx() {
-        // https://sepolia-optimism.etherscan.io/tx/0xbf8b5f08c43e4b860715cd64fc0849bbce0d0ea20a76b269e7bc8886d112fca7
-        let tx_hash: TxHash = "0xbf8b5f08c43e4b860715cd64fc0849bbce0d0ea20a76b269e7bc8886d112fca7"
-            .parse::<TxHash>()
-            .unwrap();
-
-        // https://sepolia-optimism.etherscan.io/getRawTx?tx=0xbf8b5f08c43e4b860715cd64fc0849bbce0d0ea20a76b269e7bc8886d112fca7
-        let raw_tx = alloy_primitives::hex::decode(
-            "7ef861a0dfd7ae78bf3c414cfaa77f13c0205c82eb9365e217b2daa3448c3156b69b27ac94778f2146f48179643473b82931c4cd7b8f153efd94778f2146f48179643473b82931c4cd7b8f153efd872386f26fc10000872386f26fc10000830186a08080",
-        )
-        .unwrap();
-        let dep_tx = FoundryTxEnvelope::decode(&mut raw_tx.as_slice()).unwrap();
-
-        let mut encoded = Vec::new();
-        dep_tx.encode_2718(&mut encoded);
-
-        assert_eq!(raw_tx, encoded);
-
-        assert_eq!(tx_hash, dep_tx.hash());
-    }
-
-    #[test]
     fn can_recover_sender_not_normalized() {
         let bytes = hex::decode("f85f800182520894095e7baea6a6c7c4c2dfeb977efac326af552d870a801ba048b55bfa915ac795c431978d8a6a992b628d557da5ff759b307d495a36649353a0efffd310ac743f371de3b9f7f9cb56c0b28ad43601b4ab949f53faa07bd2c804").unwrap();
 
@@ -707,11 +561,6 @@ mod tests {
         assert_eq!(tx_env.caller, sender);
         assert_eq!(tx_env.gas_limit, 0x5208);
         assert_eq!(tx_env.gas_price, 1);
-
-        // Test OpTransaction<TxEnv> conversion via FromRecoveredTx trait
-        let op_tx = OpTransaction::<TxEnv>::from_recovered_tx(&typed_tx, sender);
-        assert_eq!(op_tx.base.caller, sender);
-        assert_eq!(op_tx.base.gas_limit, 0x5208);
     }
 
     // Test vector from Tempo testnet:
