+++ crates/primitives/src/transaction/mod.rs
@@ -1,7 +1,11 @@
 mod envelope;
+#[cfg(feature = "optimism")]
+mod optimism;
 mod receipt;
 mod request;
 
 pub use envelope::{FoundryTxEnvelope, FoundryTxType, FoundryTypedTx};
+#[cfg(feature = "optimism")]
+pub use optimism::get_deposit_tx_parts;
 pub use receipt::FoundryReceiptEnvelope;
-pub use request::{FoundryTransactionRequest, get_deposit_tx_parts};
+pub use request::FoundryTransactionRequest;
