+++ crates/primitives/src/transaction/receipt.rs
@@ -8,6 +8,7 @@ use alloy_network::eip2718::{
 use alloy_primitives::{Bloom, Log, TxHash, logs_bloom};
 use alloy_rlp::{BufMut, Decodable, Encodable, Header, bytes};
 use alloy_rpc_types::{BlockNumHash, trace::otterscan::OtsReceipt};
+#[cfg(feature = "optimism")]
 use op_alloy_consensus::{
     DEPOSIT_TX_TYPE_ID, OpDepositReceipt, OpDepositReceiptWithBloom, POST_EXEC_TX_TYPE_ID,
 };
@@ -29,8 +30,10 @@ pub enum FoundryReceiptEnvelope<T = Log> {
     Eip4844(ReceiptWithBloom<Receipt<T>>),
     #[serde(rename = "0x4", alias = "0x04")]
     Eip7702(ReceiptWithBloom<Receipt<T>>),
+    #[cfg(feature = "optimism")]
     #[serde(rename = "0x7D", alias = "0x7d")]
     PostExec(ReceiptWithBloom<Receipt<T>>),
+    #[cfg(feature = "optimism")]
     #[serde(rename = "0x7E", alias = "0x7e")]
     Deposit(OpDepositReceiptWithBloom<T>),
     #[serde(rename = "0x76")]
@@ -44,7 +47,8 @@ impl FoundryReceiptEnvelope<alloy_rpc_types::Log> {
         cumulative_gas_used: u64,
         logs: impl IntoIterator<Item = alloy_rpc_types::Log>,
         tx_type: FoundryTxType,
-        deposit_nonce: Option<u64>,
+        #[cfg_attr(not(feature = "optimism"), allow(unused_variables))] deposit_nonce: Option<u64>,
+        #[cfg_attr(not(feature = "optimism"), allow(unused_variables))]
         deposit_receipt_version: Option<u64>,
     ) -> Self {
         let logs = logs.into_iter().collect::<Vec<_>>();
@@ -67,9 +71,11 @@ impl FoundryReceiptEnvelope<alloy_rpc_types::Log> {
             FoundryTxType::Eip7702 => {
                 Self::Eip7702(ReceiptWithBloom { receipt: inner_receipt, logs_bloom })
             }
+            #[cfg(feature = "optimism")]
             FoundryTxType::PostExec => {
                 Self::PostExec(ReceiptWithBloom { receipt: inner_receipt, logs_bloom })
             }
+            #[cfg(feature = "optimism")]
             FoundryTxType::Deposit => {
                 let inner = OpDepositReceiptWithBloom {
                     receipt: OpDepositReceipt {
@@ -112,13 +118,18 @@ impl FoundryReceiptEnvelope<Log> {
                 removed: false,
             })
             .collect::<Vec<_>>();
+        #[cfg(feature = "optimism")]
+        let (deposit_nonce, deposit_receipt_version) =
+            (self.deposit_nonce(), self.deposit_receipt_version());
+        #[cfg(not(feature = "optimism"))]
+        let (deposit_nonce, deposit_receipt_version) = (None, None);
         FoundryReceiptEnvelope::<alloy_rpc_types::Log>::from_parts(
             self.status(),
             self.cumulative_gas_used(),
             logs,
             self.tx_type(),
-            self.deposit_nonce(),
-            self.deposit_receipt_version(),
+            deposit_nonce,
+            deposit_receipt_version,
         )
     }
 }
@@ -132,7 +143,9 @@ impl<T> FoundryReceiptEnvelope<T> {
             Self::Eip1559(_) => FoundryTxType::Eip1559,
             Self::Eip4844(_) => FoundryTxType::Eip4844,
             Self::Eip7702(_) => FoundryTxType::Eip7702,
+            #[cfg(feature = "optimism")]
             Self::PostExec(_) => FoundryTxType::PostExec,
+            #[cfg(feature = "optimism")]
             Self::Deposit(_) => FoundryTxType::Deposit,
             Self::Tempo(_) => FoundryTxType::Tempo,
         }
@@ -158,8 +171,12 @@ impl<T> FoundryReceiptEnvelope<T> {
             Self::Eip1559(r) => FoundryReceiptEnvelope::Eip1559(r.map_logs(f)),
             Self::Eip4844(r) => FoundryReceiptEnvelope::Eip4844(r.map_logs(f)),
             Self::Eip7702(r) => FoundryReceiptEnvelope::Eip7702(r.map_logs(f)),
+            #[cfg(feature = "optimism")]
             Self::PostExec(r) => FoundryReceiptEnvelope::PostExec(r.map_logs(f)),
-            Self::Deposit(r) => FoundryReceiptEnvelope::Deposit(r.map_receipt(|r| r.map_logs(f))),
+            #[cfg(feature = "optimism")]
+            Self::Deposit(r) => FoundryReceiptEnvelope::Deposit(
+                r.map_receipt(|r: OpDepositReceipt<T>| r.map_logs(f)),
+            ),
             Self::Tempo(r) => FoundryReceiptEnvelope::Tempo(r.map_logs(f)),
         }
     }
@@ -182,35 +199,11 @@ impl<T> FoundryReceiptEnvelope<T> {
             Self::Eip1559(t) => &t.logs_bloom,
             Self::Eip4844(t) => &t.logs_bloom,
             Self::Eip7702(t) => &t.logs_bloom,
+            #[cfg(feature = "optimism")]
             Self::PostExec(t) => &t.logs_bloom,
+            #[cfg(feature = "optimism")]
             Self::Deposit(t) => &t.logs_bloom,
             Self::Tempo(t) => &t.logs_bloom,
-        }
-    }
-
-    /// Return the receipt's deposit_nonce if it is a deposit receipt.
-    pub fn deposit_nonce(&self) -> Option<u64> {
-        self.as_deposit_receipt().and_then(|r| r.deposit_nonce)
-    }
-
-    /// Return the receipt's deposit version if it is a deposit receipt.
-    pub fn deposit_receipt_version(&self) -> Option<u64> {
-        self.as_deposit_receipt().and_then(|r| r.deposit_receipt_version)
-    }
-
-    /// Returns the deposit receipt if it is a deposit receipt.
-    pub const fn as_deposit_receipt_with_bloom(&self) -> Option<&OpDepositReceiptWithBloom<T>> {
-        match self {
-            Self::Deposit(t) => Some(t),
-            _ => None,
-        }
-    }
-
-    /// Returns the deposit receipt if it is a deposit receipt.
-    pub const fn as_deposit_receipt(&self) -> Option<&OpDepositReceipt<T>> {
-        match self {
-            Self::Deposit(t) => Some(&t.receipt),
-            _ => None,
         }
     }
 
@@ -222,8 +215,10 @@ impl<T> FoundryReceiptEnvelope<T> {
             | Self::Eip1559(t)
             | Self::Eip4844(t)
             | Self::Eip7702(t)
-            | Self::PostExec(t)
             | Self::Tempo(t) => t.receipt,
+            #[cfg(feature = "optimism")]
+            Self::PostExec(t) => t.receipt,
+            #[cfg(feature = "optimism")]
             Self::Deposit(t) => t.receipt.into_inner(),
         }
     }
@@ -236,8 +231,10 @@ impl<T> FoundryReceiptEnvelope<T> {
             | Self::Eip1559(t)
             | Self::Eip4844(t)
             | Self::Eip7702(t)
-            | Self::PostExec(t)
             | Self::Tempo(t) => &t.receipt,
+            #[cfg(feature = "optimism")]
+            Self::PostExec(t) => &t.receipt,
+            #[cfg(feature = "optimism")]
             Self::Deposit(t) => &t.receipt.inner,
         }
     }
@@ -287,7 +284,9 @@ impl Encodable for FoundryReceiptEnvelope {
                     Self::Eip1559(r) => r.length() + 1,
                     Self::Eip4844(r) => r.length() + 1,
                     Self::Eip7702(r) => r.length() + 1,
+                    #[cfg(feature = "optimism")]
                     Self::PostExec(r) => r.length() + 1,
+                    #[cfg(feature = "optimism")]
                     Self::Deposit(r) => r.length() + 1,
                     Self::Tempo(r) => r.length() + 1,
                     _ => unreachable!("receipt already matched"),
@@ -314,11 +313,13 @@ impl Encodable for FoundryReceiptEnvelope {
                         EIP7702_TX_TYPE_ID.encode(out);
                         r.encode(out);
                     }
+                    #[cfg(feature = "optimism")]
                     Self::PostExec(r) => {
                         Header { list: true, payload_length: payload_len }.encode(out);
                         POST_EXEC_TX_TYPE_ID.encode(out);
                         r.encode(out);
                     }
+                    #[cfg(feature = "optimism")]
                     Self::Deposit(r) => {
                         Header { list: true, payload_length: payload_len }.encode(out);
                         DEPOSIT_TX_TYPE_ID.encode(out);
@@ -371,18 +372,23 @@ impl Decodable for FoundryReceiptEnvelope {
                     buf.advance(1);
                     <ReceiptWithBloom as Decodable>::decode(buf)
                         .map(FoundryReceiptEnvelope::Eip7702)
-                } else if receipt_type == POST_EXEC_TX_TYPE_ID {
-                    buf.advance(1);
-                    <ReceiptWithBloom as Decodable>::decode(buf)
-                        .map(FoundryReceiptEnvelope::PostExec)
-                } else if receipt_type == DEPOSIT_TX_TYPE_ID {
-                    buf.advance(1);
-                    <OpDepositReceiptWithBloom as Decodable>::decode(buf)
-                        .map(FoundryReceiptEnvelope::Deposit)
                 } else if receipt_type == TEMPO_TX_TYPE_ID {
                     buf.advance(1);
                     <ReceiptWithBloom as Decodable>::decode(buf).map(FoundryReceiptEnvelope::Tempo)
                 } else {
+                    #[cfg(feature = "optimism")]
+                    {
+                        if receipt_type == POST_EXEC_TX_TYPE_ID {
+                            buf.advance(1);
+                            return <ReceiptWithBloom as Decodable>::decode(buf)
+                                .map(FoundryReceiptEnvelope::PostExec);
+                        }
+                        if receipt_type == DEPOSIT_TX_TYPE_ID {
+                            buf.advance(1);
+                            return <OpDepositReceiptWithBloom as Decodable>::decode(buf)
+                                .map(FoundryReceiptEnvelope::Deposit);
+                        }
+                    }
                     Err(alloy_rlp::Error::Custom("invalid receipt type"))
                 }
             }
@@ -404,7 +410,9 @@ impl Typed2718 for FoundryReceiptEnvelope {
             Self::Eip1559(_) => EIP1559_TX_TYPE_ID,
             Self::Eip4844(_) => EIP4844_TX_TYPE_ID,
             Self::Eip7702(_) => EIP7702_TX_TYPE_ID,
+            #[cfg(feature = "optimism")]
             Self::PostExec(_) => POST_EXEC_TX_TYPE_ID,
+            #[cfg(feature = "optimism")]
             Self::Deposit(_) => DEPOSIT_TX_TYPE_ID,
             Self::Tempo(_) => TEMPO_TX_TYPE_ID,
         }
@@ -419,7 +427,9 @@ impl Encodable2718 for FoundryReceiptEnvelope {
             Self::Eip1559(r) => 1 + r.length(),
             Self::Eip4844(r) => 1 + r.length(),
             Self::Eip7702(r) => 1 + r.length(),
+            #[cfg(feature = "optimism")]
             Self::PostExec(r) => 1 + r.length(),
+            #[cfg(feature = "optimism")]
             Self::Deposit(r) => 1 + r.length(),
             Self::Tempo(r) => 1 + r.length(),
         }
@@ -435,8 +445,10 @@ impl Encodable2718 for FoundryReceiptEnvelope {
             | Self::Eip1559(r)
             | Self::Eip4844(r)
             | Self::Eip7702(r)
-            | Self::PostExec(r)
             | Self::Tempo(r) => r.encode(out),
+            #[cfg(feature = "optimism")]
+            Self::PostExec(r) => r.encode(out),
+            #[cfg(feature = "optimism")]
             Self::Deposit(r) => r.encode(out),
         }
     }
@@ -444,14 +456,17 @@ impl Encodable2718 for FoundryReceiptEnvelope {
 
 impl Decodable2718 for FoundryReceiptEnvelope {
     fn typed_decode(ty: u8, buf: &mut &[u8]) -> Result<Self, Eip2718Error> {
-        if ty == DEPOSIT_TX_TYPE_ID {
-            return Ok(Self::Deposit(OpDepositReceiptWithBloom::decode(buf)?));
+        #[cfg(feature = "optimism")]
+        {
+            if ty == DEPOSIT_TX_TYPE_ID {
+                return Ok(Self::Deposit(OpDepositReceiptWithBloom::decode(buf)?));
+            }
+            if ty == POST_EXEC_TX_TYPE_ID {
+                return Ok(Self::PostExec(ReceiptWithBloom::decode(buf)?));
+            }
         }
         if ty == TEMPO_TX_TYPE_ID {
             return Ok(Self::Tempo(ReceiptWithBloom::decode(buf)?));
-        }
-        if ty == POST_EXEC_TX_TYPE_ID {
-            return Ok(Self::PostExec(ReceiptWithBloom::decode(buf)?));
         }
         match ReceiptEnvelope::typed_decode(ty, buf)? {
             ReceiptEnvelope::Eip2930(tx) => Ok(Self::Eip2930(tx)),
@@ -646,8 +661,11 @@ mod tests {
         assert!(receipt.status());
         assert_eq!(receipt.cumulative_gas_used(), 100000);
         assert!(receipt.logs().is_empty());
-        assert!(receipt.deposit_nonce().is_none());
-        assert!(receipt.deposit_receipt_version().is_none());
+        #[cfg(feature = "optimism")]
+        {
+            assert!(receipt.deposit_nonce().is_none());
+            assert!(receipt.deposit_receipt_version().is_none());
+        }
     }
 
     #[test]
