+++ crates/primitives/src/transaction/request.rs
@@ -3,15 +3,19 @@ use alloy_network::{
     BuildResult, NetworkTransactionBuilder, NetworkWallet, TransactionBuilder,
     TransactionBuilder4844, TransactionBuilderError,
 };
-use alloy_primitives::{Address, B256, ChainId, TxKind, U256};
+use alloy_primitives::{Address, ChainId, TxKind, U256};
 use alloy_rpc_types::{AccessList, TransactionInputKind, TransactionRequest};
 use alloy_serde::{OtherFields, WithOtherFields};
+#[cfg(feature = "optimism")]
 use op_alloy_consensus::{DEPOSIT_TX_TYPE_ID, POST_EXEC_TX_TYPE_ID, TxDeposit};
+#[cfg(feature = "optimism")]
 use op_revm::transaction::deposit::DepositTransactionParts;
 use serde::{Deserialize, Serialize};
 use tempo_alloy::rpc::TempoTransactionRequest;
 use tempo_primitives::{TEMPO_TX_TYPE_ID, TempoTxType};
 
+#[cfg(feature = "optimism")]
+use super::optimism::get_deposit_tx_parts;
 use super::{FoundryTxEnvelope, FoundryTxType, FoundryTypedTx};
 use crate::FoundryNetwork;
 
@@ -28,6 +32,7 @@ use crate::FoundryNetwork;
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum FoundryTransactionRequest {
     Ethereum(TransactionRequest),
+    #[cfg(feature = "optimism")]
     Op(WithOtherFields<TransactionRequest>),
     Tempo(Box<TempoTransactionRequest>),
 }
@@ -44,6 +49,7 @@ impl FoundryTransactionRequest {
     pub fn into_inner(self) -> TransactionRequest {
         match self {
             Self::Ethereum(tx) => tx,
+            #[cfg(feature = "optimism")]
             Self::Op(tx) => tx.inner,
             Self::Tempo(tx) => tx.inner,
         }
@@ -55,6 +61,7 @@ impl FoundryTransactionRequest {
     /// # Returns
     /// - Ok(deposit_tx_parts) if all necessary keys are present to build a deposit transaction.
     /// - Err(missing) if some keys are missing to build a deposit transaction.
+    #[cfg(feature = "optimism")]
     pub fn get_deposit_tx_parts(&self) -> Result<DepositTransactionParts, Vec<&'static str>> {
         match self {
             Self::Op(tx) => get_deposit_tx_parts(&tx.other),
@@ -69,9 +76,11 @@ impl FoundryTransactionRequest {
     pub fn preferred_type(&self) -> FoundryTxType {
         match self {
             Self::Ethereum(tx) => tx.preferred_type().into(),
+            #[cfg(feature = "optimism")]
             Self::Op(tx) if tx.inner.transaction_type == Some(POST_EXEC_TX_TYPE_ID) => {
                 FoundryTxType::PostExec
             }
+            #[cfg(feature = "optimism")]
             Self::Op(_) => FoundryTxType::Deposit,
             Self::Tempo(_) => FoundryTxType::Tempo,
         }
@@ -95,6 +104,7 @@ impl FoundryTransactionRequest {
 
     /// Check if all necessary keys are present to build a Deposit transaction, returning a list of
     /// keys that are missing.
+    #[cfg(feature = "optimism")]
     pub fn complete_deposit(&self) -> Result<(), Vec<&'static str>> {
         self.get_deposit_tx_parts().map(|_| ())
     }
@@ -123,7 +133,9 @@ impl FoundryTransactionRequest {
             FoundryTxType::Eip1559 => self.as_ref().complete_1559(),
             FoundryTxType::Eip4844 => self.complete_4844(),
             FoundryTxType::Eip7702 => self.as_ref().complete_7702(),
+            #[cfg(feature = "optimism")]
             FoundryTxType::Deposit => self.complete_deposit(),
+            #[cfg(feature = "optimism")]
             FoundryTxType::PostExec => Err(vec!["not implemented for post-exec tx"]),
             FoundryTxType::Tempo => self.complete_tempo(),
         } {
@@ -138,9 +150,10 @@ impl FoundryTransactionRequest {
     /// Converts the request into a `FoundryTypedTx`, handling all Ethereum and OP-stack transaction
     /// types.
     pub fn build_typed_tx(self) -> Result<FoundryTypedTx, Self> {
+        #[cfg(feature = "optimism")]
         if let Ok(deposit_tx_parts) = self.get_deposit_tx_parts() {
             // Build deposit transaction
-            Ok(FoundryTypedTx::Deposit(TxDeposit {
+            return Ok(FoundryTypedTx::Deposit(TxDeposit {
                 from: self.from().unwrap_or_default(),
                 source_hash: deposit_tx_parts.source_hash,
                 to: self.kind().unwrap_or_default(),
@@ -149,8 +162,9 @@ impl FoundryTransactionRequest {
                 gas_limit: self.gas_limit().unwrap_or_default(),
                 is_system_transaction: deposit_tx_parts.is_system_transaction,
                 input: self.input().cloned().unwrap_or_default(),
-            }))
-        } else if self.complete_tempo().is_ok()
+            }));
+        }
+        if self.complete_tempo().is_ok()
             && let Self::Tempo(tx_req) = self
         {
             // Build Tempo transaction
@@ -192,6 +206,7 @@ impl Serialize for FoundryTransactionRequest {
     {
         match self {
             Self::Ethereum(tx) => tx.serialize(serializer),
+            #[cfg(feature = "optimism")]
             Self::Op(tx) => tx.serialize(serializer),
             Self::Tempo(tx) => tx.serialize(serializer),
         }
@@ -211,6 +226,7 @@ impl AsRef<TransactionRequest> for FoundryTransactionRequest {
     fn as_ref(&self) -> &TransactionRequest {
         match self {
             Self::Ethereum(tx) => tx,
+            #[cfg(feature = "optimism")]
             Self::Op(tx) => tx,
             Self::Tempo(tx) => tx.as_ref(),
         }
@@ -221,6 +237,7 @@ impl AsMut<TransactionRequest> for FoundryTransactionRequest {
     fn as_mut(&mut self) -> &mut TransactionRequest {
         match self {
             Self::Ethereum(tx) => tx,
+            #[cfg(feature = "optimism")]
             Self::Op(tx) => tx,
             Self::Tempo(tx) => tx.as_mut(),
         }
@@ -244,15 +261,16 @@ impl From<WithOtherFields<TransactionRequest>> for FoundryTransactionRequest {
             {
                 tempo_tx_req.set_nonce_key(nonce_key);
             }
-            Self::Tempo(Box::new(tempo_tx_req))
-        } else if tx.transaction_type == Some(DEPOSIT_TX_TYPE_ID)
+            return Self::Tempo(Box::new(tempo_tx_req));
+        }
+        #[cfg(feature = "optimism")]
+        if tx.transaction_type == Some(DEPOSIT_TX_TYPE_ID)
             || tx.transaction_type == Some(POST_EXEC_TX_TYPE_ID)
             || get_deposit_tx_parts(&tx.other).is_ok()
         {
-            Self::Op(tx)
-        } else {
-            Self::Ethereum(tx.into_inner())
+            return Self::Op(tx);
         }
+        Self::Ethereum(tx.into_inner())
     }
 }
 
@@ -264,6 +282,7 @@ impl From<FoundryTypedTx> for FoundryTransactionRequest {
             FoundryTypedTx::Eip1559(tx) => Self::Ethereum(Into::<TransactionRequest>::into(tx)),
             FoundryTypedTx::Eip4844(tx) => Self::Ethereum(Into::<TransactionRequest>::into(tx)),
             FoundryTypedTx::Eip7702(tx) => Self::Ethereum(Into::<TransactionRequest>::into(tx)),
+            #[cfg(feature = "optimism")]
             FoundryTypedTx::Deposit(tx) => {
                 let other = OtherFields::from_iter([
                     ("sourceHash", tx.source_hash.to_string().into()),
@@ -272,6 +291,7 @@ impl From<FoundryTypedTx> for FoundryTransactionRequest {
                 ]);
                 WithOtherFields { inner: Into::<TransactionRequest>::into(tx), other }.into()
             }
+            #[cfg(feature = "optimism")]
             FoundryTypedTx::PostExec(tx) => WithOtherFields {
                 inner: Into::<TransactionRequest>::into(tx),
                 other: OtherFields::default(),
@@ -307,8 +327,9 @@ impl From<FoundryTxEnvelope> for FoundryTransactionRequest {
     }
 }
 
-impl From<op_alloy_rpc_types::Transaction<FoundryTxEnvelope>> for FoundryTransactionRequest {
-    fn from(tx: op_alloy_rpc_types::Transaction<FoundryTxEnvelope>) -> Self {
+#[cfg(not(feature = "optimism"))]
+impl From<alloy_rpc_types_eth::Transaction<FoundryTxEnvelope>> for FoundryTransactionRequest {
+    fn from(tx: alloy_rpc_types_eth::Transaction<FoundryTxEnvelope>) -> Self {
         tx.inner.into_inner().into()
     }
 }
@@ -437,7 +458,9 @@ impl NetworkTransactionBuilder<FoundryNetwork> for FoundryTransactionRequest {
             FoundryTxType::Eip1559 => self.as_ref().complete_1559(),
             FoundryTxType::Eip4844 => self.as_ref().complete_4844(),
             FoundryTxType::Eip7702 => self.as_ref().complete_7702(),
+            #[cfg(feature = "optimism")]
             FoundryTxType::Deposit => self.complete_deposit(),
+            #[cfg(feature = "optimism")]
             FoundryTxType::PostExec => Err(vec!["not implemented for post-exec tx"]),
             FoundryTxType::Tempo => self.complete_tempo(),
         }
@@ -448,9 +471,14 @@ impl NetworkTransactionBuilder<FoundryNetwork> for FoundryTransactionRequest {
     }
 
     fn can_build(&self) -> bool {
-        self.as_ref().can_build()
-            || self.complete_deposit().is_ok()
-            || self.complete_tempo().is_ok()
+        if self.as_ref().can_build() || self.complete_tempo().is_ok() {
+            return true;
+        }
+        #[cfg(feature = "optimism")]
+        if self.complete_deposit().is_ok() {
+            return true;
+        }
+        false
     }
 
     fn output_tx_type(&self) -> FoundryTxType {
@@ -465,7 +493,9 @@ impl NetworkTransactionBuilder<FoundryNetwork> for FoundryTransactionRequest {
             FoundryTxType::Eip1559 => self.as_ref().complete_1559().ok(),
             FoundryTxType::Eip4844 => self.as_ref().complete_4844().ok(),
             FoundryTxType::Eip7702 => self.as_ref().complete_7702().ok(),
+            #[cfg(feature = "optimism")]
             FoundryTxType::Deposit => self.complete_deposit().ok(),
+            #[cfg(feature = "optimism")]
             FoundryTxType::PostExec => self.complete_type(pref).ok(),
             FoundryTxType::Tempo => self.complete_tempo().ok(),
         }?;
@@ -479,11 +509,21 @@ impl NetworkTransactionBuilder<FoundryNetwork> for FoundryTransactionRequest {
         let inner = self.as_mut();
         inner.transaction_type = Some(preferred_type as u8);
         inner.gas.is_none().then(|| inner.set_gas_limit(Default::default()));
-        if !matches!(preferred_type, FoundryTxType::Deposit | FoundryTxType::Tempo) {
+        let is_deposit = {
+            #[cfg(feature = "optimism")]
+            {
+                preferred_type == FoundryTxType::Deposit
+            }
+            #[cfg(not(feature = "optimism"))]
+            {
+                false
+            }
+        };
+        if !is_deposit && preferred_type != FoundryTxType::Tempo {
             inner.trim_conflicting_keys();
             inner.populate_blob_hashes();
         }
-        if preferred_type != FoundryTxType::Deposit {
+        if !is_deposit {
             inner.nonce.is_none().then(|| inner.set_nonce(Default::default()));
         }
         if matches!(preferred_type, FoundryTxType::Legacy | FoundryTxType::Eip2930) {
@@ -548,42 +588,10 @@ impl TransactionBuilder4844 for FoundryTransactionRequest {
     }
 }
 
-/// Converts `OtherFields` to `DepositTransactionParts`, produces error with missing fields
-pub fn get_deposit_tx_parts(
-    other: &OtherFields,
-) -> Result<DepositTransactionParts, Vec<&'static str>> {
-    let mut missing = Vec::new();
-    let source_hash =
-        other.get_deserialized::<B256>("sourceHash").transpose().ok().flatten().unwrap_or_else(
-            || {
-                missing.push("sourceHash");
-                Default::default()
-            },
-        );
-    let mint = other
-        .get_deserialized::<U256>("mint")
-        .transpose()
-        .unwrap_or_else(|_| {
-            missing.push("mint");
-            Default::default()
-        })
-        .map(|value| value.to::<u128>());
-    let is_system_transaction =
-        other.get_deserialized::<bool>("isSystemTx").transpose().ok().flatten().unwrap_or_else(
-            || {
-                missing.push("isSystemTx");
-                Default::default()
-            },
-        );
-    if missing.is_empty() {
-        Ok(DepositTransactionParts { source_hash, mint, is_system_transaction })
-    } else {
-        Err(missing)
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use alloy_primitives::B256;
+
     use super::*;
 
     fn default_tx_req() -> TransactionRequest {
@@ -618,6 +626,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "optimism")]
     fn test_routing_op_by_deposit_fields() {
         let tx = default_tx_req();
         let mut other = OtherFields::default();
@@ -669,6 +678,7 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "optimism")]
     fn test_serialization_op() {
         let tx = default_tx_req();
         let mut other = OtherFields::default();
